summary_reviews/requirements.txt
openai
langchain
langchain_core
langchain_openai
googlemaps
streamlit
python-dotenv
flask
requests